scene load refactoring
