+++ UnityProject/GravitableEscape/Assets/Script/SceneManaging/SceneFadeManager.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 
 public class SceneFadeManager : MonoBehaviour
